feat(server): Extrapolate sampled metrics numbers (#3753)
Extrapolates numbers tracked as metrics after extraction from spans.
This allows to view the original order of magnitude before client
sampling and should remove jumps resulting from changes in sample rates.

Extrapolation is part of metric extraction in Relay and controlled via a
project config field. In addition to that, there is a global threshold
for the maximum duplication factor for expansion of distribution
metrics.