ref(metrics): Flatten aggregator config (#3783)
`AggregatorServiceConfig` duplicates all of `AggregatorConfig` fields
unnecessarily. Use `#[serde(flatten)]` to deduplicate.

The idea behind having two different types was that `Aggregator` should
not get config fields that are a concern of `AggregatorService`. Since
this affects only two fields, the alternative would be to remove
`AggregatorServiceConfig` and put everything in `AggregatorConfig`
again.