fix(spans): Fixes span outcomes and inherited rate limits (#3793)
Span outcomes are currently off when created from rate limited
transactions in the `transaction_indexed` category. All of the dropped
outcomes are also added to the accepted, because the metrics are not
dropped.

Additionally outcomes for the span category were emitted when there was
a limit for `transaction_indexed` instead of just emitting them from
`span_indexed`.

Implementation notes:
Need to drop the metrics based on the enforcement instead of the rate
limits, because it is correct that there are no rate limits added for
the inherited category. We wouldn't want to start dropping standalone
spans because of a transaction quota.