fix(metrics): remove Redis spans from Queries insights module
Redis spans were intended to be filtered out of the Queries insights module,
since it currently only supports SQL. This used to work via a check for
`*redis*` in the span op. With the JS SDK's v8 release the span op for redis
spans became simply `db` and Redis spans starting showing up in the Queries module.

JS v8 follows the OTel standard of setting the `db.system` attribute to
`"redis"`, so we can use that as a secondary check to bring back the expected
filtering.