feat(performance): Update performance score calculation to checking profiles after the first match (#3804)
Updates the performance score calculation to break after find a profile
with a matching condition. This prevents any subsequent matching
profiles from overwriting calculated performance scores. This is needed
because we will be adding a `default` fallback profile with a condition
matching on any to the bottom of the profiles list.

#skip-changelog