ref(project): Treat invalid project states as pending and unify state types (#3770)
We currently deal with different envelope states inconsistently: Because
the state of a project is encoded into many separate fields and types,
it's hard to enforce that the config is validated correctly everywhere
that it is used.

For example, invalid (i.e. unparsable) project configs are sometimes
treated the same as pending, sometimes the same as disabled.