ref(processor): Switch processor to a rayon thread pool (#3833)
Switches the processor from `spawn_blocking` to a rayon thread pool. We
don't need to use the Tokio runtime for the worker threads we can use a
dedicated thread pool for this.

The problem with using the tokio thread pool is, the tokio pool may grow
largely through other uses of `tokio::spawn_blocking`, the processor
will re-use these threads and will re-initialize all thread local caches
on these threads as well. This is a large memory waste and inefficiency.