feat(metrics): Support booleans in metric extraction (#3875)
Allows to specify boolean fields as source for metric tag values in
generic metric extraction. The values will be extracted as capitalized
`"True"` and `"False"` to match how tag values are converted by
`LenientString` in event payloads.

Numbers are still explicitly unsupported as they are a major footgun
with regards to cardinality.