feat(spool): Introduce in-memory `EnvelopeBuffer` as alternative to the old spooler (#3863)
This PR introduces a new envelope buffer as designed in
https://github.com/getsentry/team-ingest/issues/345, though which all
envelopes flow before being processed. Buffered envelopes are stored in
per-project stacks. Stacks are sorted by whether the project configs
needed for processing are ready, and by latest receive time, such that
the latest envelope with a valid project config is always the first in
line to be processed.

Currently, only memory-based envelope stacks are supported, but this PR
already contains the implementation for disk-based stacks as well.

---------

Co-authored-by: Riccardo Busetti <riccardo.busetti@sentry.io>
Co-authored-by: Sebastian Zivota <loewenheim@users.noreply.github.com>