feat(process-value): Add object level trimming (#3877)
This PR adds an object-level trimming control, which works slightly
differently from before. Instead of avoiding trimming altogether, this
implementation performs all the traversal as if it were trimming but
doesn't perform the actual trimming. This way, any arbitrary node in the
traversal path can treat a child node with `trim = disabled` as if it
were trimmed, but its contents in reality are not.

In our example, we want to limit an array of spans to 800 KiB, but we
don't want the trimming to happen inside the span. Rather, we want to
let the system trim the entire spans in the array. To achieve this, we
had to implement this system in which a span signals how many bytes were
used but no data is trimmed. Then, the array sees that the remaining
size is 0 and will start dropping entire spans.

---------

Co-authored-by: Joris Bayer <joris.bayer@sentry.io>