fix(metrics): Properly split large buckets repeatedly (#3893)
A bug in `BucketView::split` causes an endless loop when splitting a
bucket more than once. Split did not honor the index of the last split
and repeatedly returned the same view on subsequent calls.

This function is used when sending global metric partitions, which
causes high request volume and backlogs when large buckets are ingested.