fix(metrics): Extract set span metrics from numeric values (#3897)
Numbers in the data couldn't be used as sets. This came up as an issue
when configuring metric extraction rules, especially since setting a
number as a tag would implicitly convert the value to a string and make
the set work. Adds the extraction for numbers as well, but convert them
to a string first, this allows for consistent sets across tags and data
as well better behaviour for users.