ref(project-cache): Check revision before loading from Redis (#3892)
Checks the 'new' revision key before actually loading and parsing the
project config.

Revision introduced here: https://github.com/getsentry/sentry/pull/75523
Change is backwards compatible, if the key is missing the check is
simply skipped.

I was considering propagating the information whether something was
refreshed or a fresh fetch through the state channel, but after review
and consideration, I don't think that is useful information, we still
have to reset a bunch of internal state and move the project out of the
`in flight` status. All if which we already do, the only thing we could
change is skipping the assignment of the project info contained in the
project fetch state (because that's the same state as the old), but this
is already only an `Arc`.