ref(metrics): Move max aggregator size into config and allow passing now timestamp for merge (#3904)
The `max_total_bucket_bytes` config can be just part of the aggregator,
it's always passed as an argument with the same value and the project
max size is already part of the aggregator.

Additionally improves performance by not re-creating the `now` timestamp
over and over again when merging a large batch of metrics into the
aggregator. This takes up a significant amount of time.