feat(project-config): Query project configs with revision (#3947)
Follow-up to to #3892, now also implements the revision fetching in the
project config API.

Sentry does not need to be updated and this is not a breaking change and
works with any combinations of Relays.

Extends the project config API with another field `revisions` which
optionally contains the revisions for all requested public keys. If
there are are revisions, the response is extended with a field
`unchanged` which contains all project keys which were unchanged (have
the same revision), similar to how `pending` is handled.