tmp(statsd): Disable high-volume timers to assess performance (#3955)
In recent refactors we noticed an increase in CPU utilization that could
be caused by more statsd timers. These are not aggegated in memory and
sent directly to the statsd agent.

This is an experiment to assess the CPU impact of the highest volume
timers that Relay tracks. If we notice a significant jump, we will
investigate further.