ref(spool): Make EnvelopeBuffer a Service (#3965)
From the linked ticket:

> We decided not to write EnvelopeBuffer as a service because we did not
want to put an additional queue in front of it (every service has an
inbound queue).
>
> However, since we have multiple request handler threads, we need some
kind of synchronization when pushing to the buffer, so in the end we
always end up with contention i.e. queueing, whether we queue threads,
tokio tasks, or messages. The current implementation spawns a tokio task
for every push and synchronizes the buffer with a tokio mutex).
>
> The current solution is in no way better than having a service, so we
might as well wrap the envelope buffer in a service and remove the
tokio::spawn and mutex.