chore(dashboard): Remove the dashboard (#3970)
The dashboard was introduced to show Relay's internal metrics and logs
in a web browser to aid in development. Since its introduction, it has
not been used and maintained however. Some of the dependencies in their
current version will soon stop compiling based on warnings emitted by
the most recent Rust version.

To trim down Relay's code base and speed up builds, this PR removes the
dashboard in its entirety.