ref: Remove generate-schema command and JSON schema (#3974)
As it was not used in an automated context, we're removing the generated
JSON schema for the event protocol. This resolves inconsistencies in the
documentation of our event protocol and also improves our compile times
through fewer derives.

We will subsequently improve the documentation of Relays types and work
on a replacement mechanism to publish docs for these types. In the
meanwhile, documentation can be viewed here:
https://getsentry.github.io/relay/relay_event_schema/protocol/index.html