fixreplays): Remove segment limit (#3984)
There is a "bug" in the SDK where a replay will emit more than one
segment every 5 seconds. This leads to certain customers exceeding the
segment limit earlier than others. By removing the limit here we allow
the SDK to determine the cap based on the duration of the replay.

As a service protection measure this limit fails to catch bad actors and
only triggers when a malformed SDK exceeds its limit. Removing this
limit and logging will allow us to fix (or not) SDK issues without
impacting customers.

#skip-changelog