feat(spans): initial MongoDB description scrubbing support (#3912)
We would like to support MongoDB queries in the Queries insight module
(which is currently SQL-only).

Metrics will be shown on the Queries insights page when they fulfill the
query `has:span.description span.module:db`. Without the changes in this
PR, MongoDB spans would either lack a `span.description` or
`span.module` so they were hidden from the module.

In addition, to fully support the Queries page, the `span.action` and
`span.domain` tags are also required. Without this PR, MongoDB spans'
`span.action` was being set if it was present as `db.operation` in span
`data`, but `span.domain` was not set.

This PR:

- Allows MongoDB spans to be tagged with `span.module`: `db`
- Adds MongoDB query scrubbing and writes the scrubbed query to
`span.description`, conditional on the
`organizations:performance-queries-mongodb-extraction` feature flag
(added [here](https://github.com/getsentry/sentry/pull/74498))
- Writes the collection name to `span.domain` for MongoDB spans

Because the new query scrubbing is behind a feature flag, orgs without
it will continue to have no `span.description` and therefore continue to
not show up in the Queries module.

The short-term plan is to turn this on for a single test org to iterate
on the frontend experience, and also begin to test the cardinality of
the query scrubbing.

The scrubbing keeps the operation/collection key/value pair but
otherwise replaces leaf values in the query JSON with `"?"`. This might
need to be tuned further if there are cases where people are likely to
use high-cardinality keys.

Valid MongoDB spans must:
- contain a JSON-encoded query in their `description`
- have a `db.system` of `"mongodb"`
- have either `db.collection.name` or `db.mongodb.collection` set
- have `db.operation` set

Our JavaScript v8 SDK fulfills these criteria for MongoDB spans, and I
believe the latest Python does too (but haven't personally checked).

---------

Co-authored-by: Ash <0Calories@users.noreply.github.com>
Co-authored-by: David Herberth <david.herberth@sentry.io>