feat(spans): Extract standalone CLS span metrics and performance score (#3988)
The JS SDK allows sending CLS as a standalone span to Relay, similar
to INP standalone spans. This update enables the extraction of
standalone CLS spans to indexed and metrics datasets, with CLS
Performance score calculation. These metrics are still written to the
`transaction` namespace to not break any product areas using CLS

#skip-changelog