ref(server): Use multer directly (#3978)
Switches to direct use of `multer` instead of `axum`'s wrapper. This
allows us to use constraints to configure the maximum body size and
maximum field size directly on the `Multipart` instance instead of
handling this manually.

In order to keep using the multipart type as an extractor, we define
`Remote` which is a shallow transparent wrapper around any remote type
we would like to implement `FromRequest`, `FromRequestParts`, or
`IntoResponse` for. This way, we can use the `Multipart` and
`multer::Error` types directly in our signatures. Usage of this type is
documented on the type.