fix(spans): check sdk platform for user.geo.subregion tag (#4023)
This is an extension of this PR
https://github.com/getsentry/relay/pull/4013

I assumed that Backend SDK transactions would not have a browser
context, so i used that as a means to check if the project is a frontend
project. Turns out, python SDKs still set browser context on the
transaction (I guess this is the browser the user made the request
from?). So to fix this, we can just check the platform directly as
frontend is typically/always javascript.

---------

Co-authored-by: Joris Bayer <joris.bayer@sentry.io>