fix(buffer): Always drop stale envelopes (#4034)
When we transformed the buffer into a service, I left the time-based
eviction in the project cache. This is a bug, because it means that old
envelopes are only dropped if we managed to get their project state(s).

With this PR, stale envelopes are always evicted once they are `peek`ed.