fix(buffer): Optimistically initialize stacks as ready (#4046)
When a new stack is created, the buffer does not know the corresponding
project state(s). It previously assumed "not ready", which means it sent
a `UpdateState` message to the project state on the first envelope.

In practice, the vast majority is created when a project state is ready
and the stack was previously removed from the priority queue because it
was empty. This PR flips the default to "ready", which means that on
startup the project cache will return `NotReady` once.