fix(buffer): Reject envelopes at high watermark, unspool at low watermark (#4075)
The new envelope buffer model assumed that we did not need to check
memory in the request handler, because envelopes "would go straight to
disk anyway". In practice this does not work because even the disk
buffer uses memory to cache batches.

This PR

* Reintroduces the memory check in the request handler, rejecting
envelopes when the 95% watermark has surpassed.
* To prevent flip-flopping, introduces a lower 90% watermark for
dequeuing (was previously 95%).