fix(server): Disallow static mode for processing (#4087)
We only support processing in managed mode, so we should make it a hard
requirement for starting Relay.

Static and proxy mode only exists to put a proxy close to the SDKs in
order to reduce latency and/or provide PII-scrubbing on premise before
sending data to sentry. From [Sentry's product
docs](https://docs.sentry.io/product/relay/):

> Sentry Relay offers enterprise-grade data security by providing a
standalone service that acts as a middle layer between your application
and sentry.io.

These modes were never intended to run in combination with `processing:
true`, and doing so resulted in undefined behavior.