ref(server): Organize project services in nested modules (#4139)
Organizing all the `project_*.rs` files we have in the `services/`
module in a module structure.

```
projects
├── cache.rs // <- services/project_cache.rs
├── mod.rs
├── project
│   ├── mod.rs // <- services/project.rs
│   └── state // <- services/project/state/*
│       ├── fetch_state.rs
│       ├── info.rs
│       └── mod.rs
└── source
    ├── local.rs // <- services/project_local.rs
    ├── mod.rs
    ├── redis.rs // <- services/project_redis.rs
    └── upstream.rs // <- services/project_upstream.rs
```

This only moves files around, no functional changes. It is a preparation
to split out more of the convoluted services into smaller digestible
parts. For example, the `project_cache.rs` module contains a lot of
utility and helper types not relevant for the service itself, it should
be possible to split this out eventually.

I chose `projects` as the base module to avoid having a
`project/project/` module, also I think it makes sense, it manages all
projects and the nested module just represents a single project.