instr(buffer): Measure envelope size (#4153)
Collect a histogram metric for the size of the envelopes that get pushed
into the envelope buffer. This will help us tune batch sizes for
writing.

We already have a metric for item sizes in the request handler, but we
want one for the entire envelope & restrict it to envelopes that
actually make it to the buffer (exclude rate limited).