feat(metrics): Add target_project_id tag to root counter metrics (#4170)
Adds a new tag named `target_project_id` to the `count_per_root_project`
metrics in the spans and transactions namespace. The tag contains the
numeric identifier of the project where the transaction or spans are
sent, in contrast to the root project where the trace originated.

This tag is added so that we can track and display stats about the
individual data volume that flows from a root project into a specific
target project.

The tag key will not be added to static strings in the metrics indexer.
Therefore, it will not be possible run cross-org queries that filter or
group by this tag.