ref(mongodb): Change grouping strategy for arrays and `buffer` (#4186)
Couple of changes here:

- Arrays in MongoDB queries were problematic for grouping, because any
variant of the same query that has a different number of elements would
result in an entirely **new group** being created. This resulted in high
cardinality for a few orgs
- Removes the 'buffer' key-value pair from queries, since this is an
under the hood mechanism which has very little relevance to the query
itself. This causes unnecessary noise and makes queries hard to read, so
we can scrub them out entirely

### Examples


![image](https://github.com/user-attachments/assets/d360e760-413e-424e-b5f6-85b70684844b)

This would become:
```
"_id": {}
```


![image](https://github.com/user-attachments/assets/93c73cfc-a6b5-4a56-bc85-e858c31dbe21)
This would become:
```
"filter": {
	"_id": {
		"$in": ["..."]
	}
}
```