ref(project-cache): Schedule updates instead of spawning tasks (#4233)
We can do better than spawning a large amount of tasks which just wait
by using a heap to keep track of when the next task is scheduled and
`FuturesUnordered` to wait for the results from the tasks. Each task is
dispatched to a `ProjectSource` which is very heavily IO bound which and
we don't need parallelism for.