ref(project-cache): Schedule evictions (#4241)
The hunt for tail latencies continues.

Project cache evictions for large project caches are quite costly, we're
seeing a max latency for evictions of > 50ms. Timing the evictions of
each project is not just much faster (log(n) inserts instead of full
scans through projects), but also much shorter individually, giving the
runtime opportunities to schedule other tasks.

Implementation is largely the same as in #4233, just backed by a
`PriorityQueue` instead of a `BinaryHeap`, the future scheduling code is
shared and not changed.