ref(relay): Fix Rust 1.83 lints (#4307)
Fixes all Clippy and rustfmt issues with 1.83.

This adds a bunch of ignores for the `non_local_definitions` lint, since
`synstructure` generates code which causes the lint to fire.
`synstructure` 0.13 has this fixed, but it depends on syn2, where we
still use syn1 in the derive crates. Updating to syn2 is too much effort
for the moment.