fix(projects): Treat fetch failures as pending (#4140)
When we fail to fetch a project state for any reason (max retries
exceeded, redis error), we should not mark the project as invalid.

Currently we treat the project as if the DSN was invalid, and respond to
clients with a 403 response. This might have made sense as a defensive
measure in a pre-spooling world where envelopes could not be buffered
for a longer period of time, but with spooling and time-based envelope
eviction, we can now safely buffer them.