chore(inbound-filters): Update chrome inbound filter (#4381)
this pr updates the chrome inbound filter to filter a higher version.
goal when we added these filters last year was to update them
periodically. as of now, chrome is the only one we think we should
update. one of the resources we used last year to determine which
versions to filter was NextJS supported browsers and the Chrome version
has been updated there, so we want to make the change for the filter