ref: use version of statsdproxy that can work with Duration (#4460)
This PR uses a fixed commit for `statsdproxy` that takes a `Duration`
instead of u64 seconds as `flush_interval` so we can see if reducing the
`flush_interval` helps with the metrics problem.

This is meant as a proof of concept, if it works we will update the
`statsdproxy` crate to support it regulary.

statsdproxy PR: https://github.com/getsentry/statsdproxy/pull/55