fix(insights): strip sql comments without trailing newline (#4493)
Found this case when investigating projects emitting high cardinality
metrics. SQL comments that did not end in a newline aren't being
stripped correctly. Updated the comment-matching regex to additionally
match comments up to the end of the haystack.

Note that there's an interesting interaction with the regexes that meant
that the test
```rust
    scrub_sql_test!(
        comment_without_ending_line_break,
        "SELECT * FROM comments LIMIT 1 -- comment", // note the explicit limit
        "SELECT * FROM comments LIMIT %s"
    );
```
already passed without these changes, while the version with the named
limit param (`LIMIT %(param_1)s`) did not.