feat(cogs): Track categories in COGS (#4496)
Makes it possible to split a COGS measurement into multiple categories.

Also instruments the transaction processing with more categories. The
initial plan was to instrument normalization in multiple smaller
categories, I decided against that in this PR because that requires an
API change with huge effects in tests. If we realize normalization takes
up a significant amount, we can revisit this.

Ref: https://github.com/getsentry/team-ingest/issues/637