feat(spans): Add and populate an `is_remote` flag on spans (#4499)
Adds a top-level tri-boolean flag called `is_remote` that indicates
whether the span's parent is in a remote service. This will be used by
the future span pipeline to infer segment spans.

The semantics for this flag are taken from the OpenTelemetry spec at:
https://opentelemetry.io/docs/specs/otel/trace/api/#isremote

In the internal schema, the flag is a top-level boolean that complements
the existing `is_segment` flag. In OpenTelemetry, it is part of the
`flags` bitfield and can be parsed using OTEL-provided masks. During
span extraction, Relay sets the field like so:

- Transactions are always marked as `is_remote: true`. This makes it
  reliable to detect segments in the pipeline.
- Spans extracted from transactions are always marked with
  `is_remote: false`.
- Spans converted from otel read the ternary flag from flags based on
  `HAS_IS_REMOTE` and `IS_REMOTE`.
- By default, `is_remote` is set to `empty`.