fix(otlp): update op/description mapping
Update incoming OTLP span `name` to write to spans' `op` field instead of
`description`. If a `sentry.description` attribute is given, write that to the
spans' `description` field.

The intended product behaviour is:
- product functionality that depends on specific `op` values/formatting will be
  removed (replaced with `category`, which already exists but will be generated
  from span attributes instead of `op` - PR to follow)
- with that removed, `op` and `name` have the same semantics and use cases and
  will be made equivalent
- `description` will remain as it is today, to either be passed explicitly as
  `sentry.description` in attributes or generated by the system (PR to follow)

Sentry's POTel SDKs currently map `name` to `description` in the SDK. This will
eventually be changed to be in line with the above logic when they update to an
OTLP span-based transport.