feat(system): Instrument service utilization (#4501)
The goal: Generically track a utilization metric (and more) for each
service spawned in Relay.

To make this possible `relay-system`'s `Runtime` and `Handle` have
gained the possibility to spawn services. Now instead of instantiating a
`ServiceRunner` services are spawned through the runtime.

The runtime has now a registry for all spawned services, which allows
enumerating all currently running services as well as querying their
tracked metrics.

To make testing not abysmal, a `ServiceSpawn` trait was introduced,
tests can spawn services on the Tokio Runtime using `TokioServiceSpawn`.

This is also a great foundation to finally move the graceful shutdown
handling properly into the service framework instead of relying on a
global.