feat(otlp): write span kind to attributes
OTLP spans have a property called `kind` that helps specify the relationship
between that span and its parent or child. Sentry users should be able to see
the span kind and search/filter on it when it is sent. During ingestion, write
a string representation of the `kind` to the `span.kind` attribute.

(A string representation is chosen over the default integer one because it is
both self-describing and also makes search straightforward (no need for
translations between string and integer in the frontend or EAP. We are relying
on Clickhouse's compression to make the additional data usage reasonable.)

[0] https://github.com/open-telemetry/opentelemetry-proto/blob/d7770822d70c7bd47a6891fc9faacc66fc4af3d3/opentelemetry/proto/trace/v1/trace.proto#L152-L178

Co-authored-by: Ash Anand <ash.anand@sentry.io>