fix(otlp): update op/description mapping (#4505)
Update incoming OTLP span `name` to write to spans' `op` field instead
of `description`. If a `sentry.description` attribute is given, write
that to the spans' `description` field.

For background, this change will align with the rest of the product by
the time this OTLP endpoint is shipped:
- product functionality that depends on specific `op` values/formatting
will be removed (replaced with `category`, which already exists as a
Sentry tag but will be generated from span attributes instead of `op` -
PR to follow)
- with that removed, `op` and `name` will have the same semantics and
use cases and will be used interchangeably as "name" across the product
- functionality that depends on particular `description` values being
set (e.g. queries for `db` spans) will be removed and replaced with
references to explicit attributes (e.g. `db.query.text`)
- with that done, `description` will remain as extra context to augment
span display in the trace view/explore, to either be passed explicitly
as `sentry.description` in attributes or (in most cases) generated by
the system - PR to follow

Sentry's POTel SDKs currently map `name` to `description` in the SDK.
This will eventually be changed to be in line with the above logic when
they update to an OTLP span-based transport.