feat(ourlogs): Always set observed time for logs (#4559)
### Summary
Observed time nanos should be set by the collector in OTel. In the case
of Sentry the collector will always be Relay. We may want to allow
passing observed time nanos in the future, but we can address that after
we move to using EAP items consumers since they rely on 'received'
instead of observed time.