build(deps): update opentelemetry-proto to 0.28.0 (#4589)
As I discovered when trying to store incoming OTLP span links, v0.7.0 of
`opentelemetry-proto` can't deserialize the `traceId` or `spanId`
properties of span links in OTLP JSON (See upstream issue:
https://github.com/open-telemetry/opentelemetry-rust/issues/2065).

Updating to the latest `opentelemetry-proto` (0.28.0) fixes the issue,
allowing us to deserialize incoming span links sent as JSON to the OTLP
endpoint as well as in tests.

It looks like the JSON serialization for array attributes also changed
slightly with this update, affecting a test. Since no one is actually
using this endpoint yet it should be safe to update.