feat(links): write span links to Kafka
Span links were added in <PR>, but are currently only accessible as part of
transaction events. We would like to get them in EAP. The first step in that
process is to add them to serialized Kafka spans.

The integration tests confirm production into Kafka for every current path for
links to enter the system:
- in a transaction event's trace context
- in a transaction event's child spans
- in a `span` envelope item
- in an `otel_span` envelope item
- via OTLP JSON
- via OTLP gRPC

Co-authored-by: Lukas Stracke <lukas.stracke@sentry.io>
Co-authored-by: Sigrid Huemer <32902192+s1gr1d@users.noreply.github.com>