feat(spans): extract transaction breakdowns into measurements (#4600)
closes https://github.com/getsentry/relay/issues/4598

More details found in the issue above. This PR extracts transaction
breakdowns into measurements, so that they are automatically stored in
the eap dataset. (Snuba does not look at breakdowns at all, but it
automatically stores all measurements) We are moving the insights
overview pages to eap, and the product relies on these breakdown values.

---------

Co-authored-by: Riccardo Busetti <riccardob36@gmail.com>